+++ measures/src/commonMain/kotlin/spondee/measures/Cardinal.kt
@@ -1,0 +1,16 @@
+package spondee.measures
+
+/**
+ * Geographic Direction
+ */
+enum class Cardinal(val symbol: Char, internal val decimalSign: Int) {
+    North('N', 1),
+    East('E', 1),
+    South('S', -1),
+    West('W', -1)
+}
+
+/**
+ * Convert a single character to a cardinal direction.
+ */
+fun Char.toCardinal(): Cardinal = Cardinal.values().first { it.symbol == toUpperCase() }