+++ CHANGELOG.md
@@ -8,3 +8,5 @@ Changes
 
  - `Percentage` Unit
  - `Level` Unit
+ - `GeoLine` Units for direction: `Latitude` / `Longitude`
+ - `Cardinal` enum for indicating a N/S/E/W direction.